test appstate2

+++ miprometheus/models/cog/network.py
@@ -82,10 +82,10 @@ class CogModel(Model):
 		questions = self.forward_lookup2embed(questions)
 		
 		output_class = torch.zeros((images.size()[1],images.size()[0],2),requires_grad=False).type(dtype)
-		output_point = torch.zeros((images.size()[1],images.size()[0],49),requires_grad=False)
-		attention = torch.randn((images.size()[1],self.controller_output_size*2),requires_grad=False)
-		controller_state = torch.zeros((1,images.size()[1],128),requires_grad=False)
-		vstm_state = torch.zeros((images.size()[1],self.vstm_nmaps,self.vstm_shape[0],self.vstm_shape[1]),requires_grad=False)
+		output_point = torch.zeros((images.size()[1],images.size()[0],49),requires_grad=False).type(dtype)
+		attention = torch.randn((images.size()[1],self.controller_output_size*2),requires_grad=False).type(dtype)
+		controller_state = torch.zeros((1,images.size()[1],128),requires_grad=False).type(dtype)
+		vstm_state = torch.zeros((images.size()[1],self.vstm_nmaps,self.vstm_shape[0],self.vstm_shape[1]),requires_grad=False).type(dtype)
 
 		print(output_class.device)
 		print(output_point.device)
@@ -172,7 +172,7 @@ class CogModel(Model):
 
 	def forward_lookup2embed(self,questions):
 		
-		out_embed=torch.zeros((questions.size(0),self.nwords,self.words_embed_length),requires_grad=False)
+		out_embed=torch.zeros((questions.size(0),self.nwords,self.words_embed_length),requires_grad=False).type(dtype)
 		for i, sentence in enumerate(questions):
 			out_embed[i,:,:] = ( self.Embedding( sentence ))
